+++ src/config/mod.rs
@@ -61,6 +61,8 @@ pub struct Config {
     pub no_default_features: bool,
     /// Features to include in the target project build
     pub features: Vec<String>,
+    /// Unstable cargo features to use
+    pub unstable_features: Vec<String>,
     /// Build all packages in the workspace
     pub all: bool,
     /// Packages to include when building the target project
@@ -77,6 +79,14 @@ pub struct Config {
     pub release: bool,
     /// Build the tests only don't run coverage
     pub no_run: bool,
+    /// Don't update `Cargo.lock`.
+    pub locked: bool,
+    /// Don't update `Cargo.lock` or any caches.
+    pub frozen: bool,
+    /// Directory for generated artifacts
+    pub target_dir: Option<PathBuf>,
+    /// Run tarpaulin on project without accessing the network
+    pub offline: bool,
 }
 
 impl Default for Config {
@@ -102,6 +112,7 @@ impl Default for Config {
             forward_signals: false,
             no_default_features: false,
             features: vec![],
+            unstable_features: vec![],
             all: false,
             packages: vec![],
             exclude: vec![],
@@ -111,6 +122,10 @@ impl Default for Config {
             release: false,
             all_features: false,
             no_run: false,
+            locked: false,
+            frozen: false,
+            target_dir: None,
+            offline: false,
         }
     }
 }
@@ -141,7 +156,8 @@ impl<'a> From<&'a ArgMatches<'a>> for Config {
             all_features: args.is_present("all-features"),
             no_default_features: args.is_present("no-default-features"),
             features: get_list(args, "features"),
-            all: args.is_present("all"),
+            unstable_features: get_list(args, "Z"),
+            all: args.is_present("all") | args.is_present("workspace"),
             packages: get_list(args, "packages"),
             exclude: get_list(args, "exclude"),
             excluded_files: get_excluded(args),
@@ -149,6 +165,10 @@ impl<'a> From<&'a ArgMatches<'a>> for Config {
             test_timeout: get_timeout(args),
             release: args.is_present("release"),
             no_run: args.is_present("no-run"),
+            locked: args.is_present("locked"),
+            frozen: args.is_present("frozen"),
+            target_dir: get_target_dir(args),
+            offline: args.is_present("offline"),
         }
     }
 }